refactor: s3 폴더 이동(#122)
